final @inline and fix some warnings

+++ core/src/main/scala/swaydb/core/segment/format/a/block/reader/BlockReader.scala
@@ -53,7 +53,7 @@ private[reader] object BlockReader {
       (offset.size - fromPosition) >= atLeastSize
   }
 
-  def get(state: State): Int =
+  def get(state: State): Byte =
     if (state.hasMore) {
       val byte =
         state.
